refactor: isJoinning 상태 삭제
