+++ frontend/src/App.tsx
@@ -3,11 +3,11 @@ import Navbar from './components/Navbar';
 import React, { useEffect, useRef } from 'react';
 import queryClient from './api/queryClient';
 import { QueryClientProvider } from '@tanstack/react-query';
-import { useLoginBottomSheetStateStore } from './store/modalStateStore';
+import { useLoginBottomSheetStateStore } from './store/loginBottomSheetStore';
 import LoginBottomSheet from '@/components/LoginBottomSheet';
 var console;
 function App() {
-    const { isLoginBottomSheetOpen, isJoinning, setLoginBottomSheetState } = useLoginBottomSheetStateStore();
+    const { isLoginBottomSheetOpen, setLoginBottomSheetState } = useLoginBottomSheetStateStore();
     const outletRef = useRef<HTMLDivElement>(null);
     //TODO: 일시적인 배포시 console.log 제거 추가로 환경설정으로 빼줘야함ㄴ
     if (process.env.NODE_ENV === 'production') {
@@ -33,18 +33,16 @@ function App() {
                 <div ref={outletRef}>
                     <Outlet />
                 </div>
-                {!isJoinning && (
-                    <div className={`fixed z-10`}>
-                        <Navbar />
-                    </div>
-                )}
-                {!isJoinning && (
-                    <div
-                        className={`fixed z-20 w-full duration-300 ${isLoginBottomSheetOpen ? ' translate-y-72' : 'translate-y-full'}`}
-                    >
-                        <LoginBottomSheet />
-                    </div>
-                )}
+
+                <div className={`fixed z-10`}>
+                    <Navbar />
+                </div>
+
+                <div
+                    className={`fixed z-20 w-full duration-300 ${isLoginBottomSheetOpen ? ' translate-y-72' : 'translate-y-full'}`}
+                >
+                    <LoginBottomSheet />
+                </div>
             </div>
         </QueryClientProvider>
     );
